cornet: razor pages removed, program combined with startup
