--- CornetInterfaceService/CornetInterfaceService/Pages/Privacy.cshtml
@@ -1,8 +1,0 @@
-﻿@page
-@model PrivacyModel
-@{
-    ViewData["Title"] = "Privacy Policy";
-}
-<h2>@ViewData["Title"]</h2>
-
-<p>Use this page to detail your site's privacy policy.</p>
