--- CornetInterfaceService/CornetInterfaceService/Pages/_ViewStart.cshtml
@@ -1,3 +1,0 @@
-﻿@{
-    Layout = "_Layout";
-}
