--- CornetInterfaceService/CornetInterfaceService/wwwroot/lib/jquery-validation-unobtrusive/.bower.json
@@ -1,15 +1,0 @@
-{
-  "name": "jquery-validation-unobtrusive",
-  "homepage": "https://github.com/aspnet/jquery-validation-unobtrusive",
-  "version": "3.2.9",
-  "_release": "3.2.9",
-  "_resolution": {
-    "type": "version",
-    "tag": "v3.2.9",
-    "commit": "a91f5401898e125f10771c5f5f0909d8c4c82396"
-  },
-  "_source": "https://github.com/aspnet/jquery-validation-unobtrusive.git",
-  "_target": "^3.2.9",
-  "_originalSource": "jquery-validation-unobtrusive",
-  "_direct": true
-}
